+++ hadoop-ozone/ozone-manager/src/main/java/org/apache/hadoop/ozone/protocolPB/OzoneManagerRequestHandler.java
@@ -500,6 +500,10 @@ public class OzoneManagerRequestHandler implements RequestHandler {
    */
   @Override
   public void validateRequest(OMRequest omRequest) throws OMException {
+    requestParamValidation(omRequest);
+  }
+
+  public static void requestParamValidation(OMRequest omRequest) throws OMException {
     Type cmdType = omRequest.getCmdType();
     if (cmdType == null) {
       throw new OMException("CmdType is null",
